requirements.txt
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.2
flask_mail


google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.94.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0
